refactor: Introduce AppContext, change Signal/Vec render logic, rename EventType to Ev
* View logic
Use custom View:: variants for Signal and SignalVec rendering.
Also adds a RetainedView type which is used for keeping needed state.

* Context

Add a global AppContext type.
This is the basis for a future context system.

All UI code must run with a AppContext active, which is ensured by
wrapping all spawned futures in the new ContextFuture, which sets the
active context before each poll.

* Event

Rename Event to Ev, which is easier to use.
Also a preparation for a future custom EventType.